resolve ty issue
