+++ .devcontainer/post-create.sh
@@ -19,9 +19,25 @@ uv venv
 uv sync
 cd ..
 
-# Install pre-commit hooks
+# Install prek (pre-commit hook runner)
+if ! command -v prek &> /dev/null; then
+    echo "🪝 Installing prek..."
+    ARCH=$(uname -m)
+    case "$ARCH" in
+        x86_64)  TARGET="x86_64-unknown-linux-gnu" ;;
+        aarch64) TARGET="aarch64-unknown-linux-gnu" ;;
+        *)       echo "⚠️  Unsupported architecture: $ARCH, skipping prek install"; TARGET="" ;;
+    esac
+    if [ -n "$TARGET" ]; then
+        curl -sSL "https://github.com/j178/prek/releases/latest/download/prek-${TARGET}.tar.gz" | tar xz --strip-components=1 -C /tmp
+        install /tmp/prek /usr/local/bin/prek
+        rm -f /tmp/prek
+    fi
+fi
+
+# Install pre-commit hooks (--overwrite removes any legacy pre-commit hook)
 echo "🪝 Installing pre-commit hooks..."
-uvx pre-commit install
+prek install --overwrite
 
 # Copy .env.example if .env doesn't exist
 if [ ! -f api/.env ] && [ -f api/.env.example ]; then
